+++ tests/test_admin.py
@@ -13,7 +13,6 @@ from django.contrib import admin
 from django.contrib.auth import get_user_model
 from django.test import TestCase
 from mock import Mock, patch
-import six
 
 from djstripe.admin import reprocess_events, subscription_status
 from djstripe.models import Customer, Event, Subscription
@@ -31,7 +30,7 @@ class TestAdminSite(TestCase):
         Bad search field <customer__user__username> for Customer model.
         """
 
-        for _model, model_admin in six.iteritems(admin.site._registry):
+        for _model, model_admin in admin.site._registry.items():
             for search_field in getattr(model_admin, 'search_fields', []):
                 model_name = model_admin.model.__name__
                 self.assertFalse(search_field.startswith('{table_name}__'.format(